==> Notebooks/2_find_working_nodes.ipynb <==
# Znalezienie węzłów pracujących
Notatnik, który odfiltrowuje wszystkie nieinteresujące nas węzły, zostawiając przy tym tylko przebiegi tych co przetwarzały jakieś dane.

import pandas as pd
import glob
import os
import shutil

prepared_directory = "Prepared"
organised_directory = "Organised"
root = f".\..\{prepared_directory}"

files = glob.glob(f".\..\{prepared_directory}" + '/**/*.csv', recursive=True)
file_count = len(files)
print(f"znaleziono {file_count} plików.")
file_number = 0
progress = 1
progress_interval = file_count*0.05

for functionTypeDirectory in os.listdir(root):
    if os.path.isdir(os.path.join(root, functionTypeDirectory)):
        for functionDirectory in os.listdir(os.path.join(root, functionTypeDirectory)):
            for numberOfNodesDirectory in os.listdir(os.path.join(root, functionTypeDirectory, functionDirectory)):
                destinationPath =  os.path.join(root, functionTypeDirectory, functionDirectory, numberOfNodesDirectory, 'source-data')
                for nodeDirectory in os.listdir(destinationPath):
                    path =  os.path.join(destinationPath, nodeDirectory)
                    all_files = glob.glob(path + "/*.csv")
                    i = 1
                    for filename in all_files:
                        if file_number > progress_interval*progress:
                            progress_string = "#".rjust(progress, "#").ljust(20, '-')
                            print(f"{progress_string} ({file_number}/{file_count})")
                            progress += 1
                        file_number+=1
                        
                        if os.path.exists(os.path.join(destinationPath.replace(prepared_directory, organised_directory), str(i))):
                            continue
                        df = pd.read_csv(filename)
                        if df['RAM'].mean() >= 2 and df['CPU'].head(15).mean() > 3:
                            if not os.path.exists(os.path.join(destinationPath.replace(prepared_directory, organised_directory), str(i))):
                                os.makedirs(os.path.join(destinationPath.replace(prepared_directory, organised_directory), str(i)))
                            shutil.copyfile(filename, os.path.join(destinationPath.replace(prepared_directory, organised_directory), str(i), os.path.basename(filename)))
                        i += 1
print("skończone.")
