==> Notebooks/3_join_files.ipynb <==
# Połączenie plików z węzłów
Notatnik, który zbiera wszystkie pliki, gdzie każdy jest pojedyńczym przebiegiem. Nadaje im etykiete (label) i tworzy jest zbiory plik dla każdego typu udf'a.

import pandas as pd
import os

organised_directory = "Organised"
labels = ['aggregation', 'filtration', 'filtration-aggregation', 'filtration-aggregation-join', 'filtration-join']
joined_df = pd.DataFrame()

for label in labels:
    print(f"Zaczęto typ {label}")
    directory = f"./../{organised_directory}/{label}"
    udf_dirs = os.listdir(directory)
    result_df = pd.DataFrame()
    snapshot = 0
    for udf in udf_dirs:
        for root, _, files in os.walk(f"./{directory}/{udf}"):
            for file in files:
                full_path = os.path.join(root, file)

                node_data_df = pd.read_csv(full_path)
                node_data_df['snapshot'] = snapshot
                snapshot += 1
                node_data_df["label"] = label
                node_data_df["udf"] = udf
                node_data_df["size"] = full_path.split("\\")[-4].split("-")[3][0]
                result_df = pd.concat([result_df,node_data_df])
    joined_df = pd.concat([joined_df, result_df[["snapshot", "label", "udf", "epoch", "CPU", "RAM", "size"]]])
    print(f"Przetworzono typ {label}")
joined_file_name = f"./../{organised_directory}/Time_series_udf_dataset.csv"
joined_df.to_csv(joined_file_name, index = False)
print(f"Zapisano plik {joined_file_name}")
