==> Notebooks/4_normalize_data.ipynb <==
# Normalizacja danych
Notatnik służaczy do normalizacji danych metodą Min-Max. Wynik zapisany jest w folderze 'Organised' pod nazwą 'Time_series_udf_dataset_normalized.csv'

import pandas as pd
organised_directory = "Organised"

full_df = pd.read_csv(f"./../{organised_directory}/Time_series_udf_dataset.csv")
full_df.CPU = (full_df.CPU-full_df.CPU.min())/(full_df.CPU.max() - full_df.CPU.min())
full_df.RAM = (full_df.RAM-full_df.RAM.min())/(full_df.RAM.max() - full_df.RAM.min())
full_df.to_csv(f"./../{organised_directory}/Time_series_udf_dataset_normalized.csv", index = False)
